+++ visionatrix/install_update/custom_nodes.py
@@ -9,78 +9,9 @@ import httpx
 from packaging.version import Version
 
 from .. import _version, options
-from ..models import install_model
-from ..pydantic_models import AIResourceModel
+from ..basic_node_list import BASIC_NODE_LIST
 
 LOGGER = logging.getLogger("visionatrix")
-BASIC_NODE_LIST = {
-    "ComfyUI-Impact-Pack": {
-        "main_branch": "Main",
-    },
-    "ComfyUI_UltimateSDUpscale": {
-        "git_flags": "--recursive",
-    },
-    "ComfyUI_InstantID": {
-        "requirements": {
-            "insightface": {},
-            "onnxruntime": {},
-        },
-        "models": [
-            AIResourceModel(
-                name="antelopev2",
-                save_path="{root}models/insightface/models/antelopev2.zip",
-                url="https://huggingface.co/MonsterMMORPG/tools/resolve/main/antelopev2.zip",
-                hash="8e182f14fc6e80b3bfa375b33eb6cff7ee05d8ef7633e738d1c89021dcf0c5c5",
-            ),
-        ],
-    },
-    "ComfyUI-BRIA_AI-RMBG": {
-        "models": [
-            AIResourceModel(
-                name="RMGB-1.4",
-                save_path="{root}custom_nodes/ComfyUI-BRIA_AI-RMBG/RMBG-1.4/model.pth",
-                url="https://huggingface.co/andrey18106/vix_models/resolve/main/RMBG-1.4/model.pth",
-                homepage="https://huggingface.co/briaai/RMBG-1.4",
-                hash="893c16c340b1ddafc93e78457a4d94190da9b7179149f8574284c83caebf5e8c",
-            ),
-        ],
-    },
-    "efficiency-nodes-comfyui": {
-        "requirements": {
-            "simpleeval": {},
-        }
-    },
-    "ComfyUI-WD14-Tagger": {},
-    "ComfyUI-SUPIR": {},
-    "ComfyUI_essentials": {},
-    "rgthree-comfy": {},
-    "ComfyUI-Custom-Scripts": {},
-    "ComfyUI_IPAdapter_plus": {},
-    "comfyui_controlnet_aux": {},
-    "Skimmed_CFG": {},
-    "comfyui-art-venture": {},
-    "was-node-suite-comfyui": {},
-    "ComfyUI-Visionatrix": {},
-    "comfyui-ollama": {},
-    "ComfyUI-AutoCropFaces": {},
-    "PuLID_ComfyUI": {
-        "before_install": {
-            "python": "-m pip install --use-pep517 facexlib",
-        },
-        "requirements": {
-            "facexlib": {},
-            "insightface": {},
-            "onnxruntime": {},
-            "ftfy": {},
-            "timm": {},
-        },
-    },
-    "ComfyUI_FizzNodes": {},
-    "style_aligned_comfy": {
-        "main_branch": "master",
-    },
-    "ComfyUI_Gemini_Flash": {},
-}
 
 
 def install_base_custom_nodes() -> None:
@@ -110,9 +41,6 @@ def install_base_custom_node(node_name: str, node_details: dict) -> None:
     _before_install(node_name, node_details)
     _install_requirements(custom_nodes_dir, node_name, node_details)
     _run_install_script(custom_nodes_dir, node_name)
-    if "models" in node_details:
-        for i in node_details["models"]:
-            install_model(i, {}, None)
 
 
 def update_base_custom_nodes() -> None: