+++ visionatrix/install_update/update.py
@@ -6,7 +6,7 @@ from subprocess import CalledProcessError, check_call
 
 from packaging.version import Version
 
-from .. import _version, options
+from .. import _version, comfyui, options
 from ..flows import get_available_flows, get_installed_flows, install_custom_flow
 from .custom_nodes import update_base_custom_nodes
 from .install import create_missing_models_dirs
@@ -20,16 +20,6 @@ def __progress_callback(name: str, progress: float, error: str) -> None:
 
 
 def update() -> None:
-    logging.info("Updating flows..")
-    avail_flows_comfy = []
-    avail_flows = get_available_flows(avail_flows_comfy)
-    avail_flows_names = [v.name for v in avail_flows]
-    for i in get_installed_flows():
-        if i.name in avail_flows_names:
-            v = avail_flows_names.index(i.name)
-            install_custom_flow(avail_flows[v], avail_flows_comfy[v], __progress_callback)
-        else:
-            logging.warning("`%s` flow not found in repository, skipping update of it.", i.name)
     if options.BACKEND_DIR:
         if options.PYTHON_EMBEDED:
             create_missing_models_dirs()
@@ -58,3 +48,14 @@ def update() -> None:
         create_missing_models_dirs()
         logging.info("Updating custom nodes..")
         update_base_custom_nodes()
+    comfyui.load(None)
+    logging.info("Updating flows..")
+    avail_flows_comfy = []
+    avail_flows = get_available_flows(avail_flows_comfy)
+    avail_flows_names = [v.name for v in avail_flows]
+    for i in get_installed_flows():
+        if i.name in avail_flows_names:
+            v = avail_flows_names.index(i.name)
+            install_custom_flow(avail_flows[v], avail_flows_comfy[v], __progress_callback)
+        else:
+            logging.warning("`%s` flow not found in repository, skipping update of it.", i.name)